Organized MITM-proxy utils
