+++ mitm-utils/whitelist-mitm.js
